ビルトイン関数を一つのファイルに
