+++ include/stdarg.h
@@ -1,2 +1,4 @@
 typedef void *va_list;
 #define va_start __builtin_va_start
+#define va_arg __builtin_va_arg
+#define va_end __builtin_va_end
